Revamp DefaultCodegenConfig fromRequestBody()
This makes the behaviour of `fromRequestBody()` similar to that of
`fromParameter()` (by replacing most of the existing code with code
from `fromParameter()`) in order to fix a number of issues. Notably this
includes a lack of handling for enums, both as references (as can be
seen in #95) and inline.

Also:
 - Add `postProcessRequestBody()` hook for code generators
 - Fix #95