#592 Ported automatic access token retrieval from openapi-generator
The OpenAPITools/openapi-generator generates Java client SDK with okHttp which provides an automatic retrieval of an OAuth2 access token. This is done by registering an interceptor (RetryingOAuth class) extending the OAuth class to the okHTTP client instantiated in the ApiClient class. With that any outgoing HTTP call from the client SDK is intercepted and preceeding call to the Auth-Server is done first to retrieve an access token. That access token is then added to the original outgoing HTTP call providing an authorized access to the actual API.