[Python] Lazily initialize thread pool
Porting a Swagger Codegen 2 fix to Swagger Codegen 3.
The ThreadPool `pool` is only used for async requests.
If a client is not making any async requests, the pool is
initialized and needs to be shutdown regardless. Taking up
resources (threads) and putting a burden on the end user
of the generated client to properly shut down the thread
pool.

https://github.com/swagger-api/swagger-codegen/commit/ebc8313d09cae208b953a8a5ba4742cc7d38d44f#diff-d7b82d024488f96a18a1c76247f1bd57cbde141f28defe6f45b04ce566ab5aeb